+++ packages/manager/cypress/support/intercepts/linodes.ts
@@ -779,3 +779,14 @@ export const mockGetLinodeStats = (
     makeResponse(stats)
   );
 };
+
+/**
+ * Intercepts PUT request to edit details of a linode
+ *
+ * @param linodeId - ID of Linode for intercepted request.
+ *
+ * @returns Cypress chainable.
+ */
+export const mockUpdateLinode = (linodeId: number): Cypress.Chainable<null> => {
+  return cy.intercept('PUT', apiMatcher(`linode/instances/${linodeId}`));
+};