+++ packages/manager/src/features/Billing/BillingPanels/BillingSummary/BillingSummary.test.tsx
@@ -5,20 +5,12 @@ import * as React from 'react';
 
 import { PAYPAL_CLIENT_ID } from 'src/constants';
 import { promoFactory } from 'src/factories';
-import {
-  renderWithThemeAndRouter,
-  withMarkup,
-  wrapWithThemeAndRouter,
-} from 'src/utilities/testHelpers';
+import { renderWithTheme, withMarkup } from 'src/utilities/testHelpers';
 
 import BillingSummary from './BillingSummary';
 
 const accountBalanceText = 'account-balance-text';
 const accountBalanceValue = 'account-balance-value';
-
-const queryMocks = vi.hoisted(() => ({
-  useMatch: vi.fn().mockReturnValue({}),
-}));
 
 vi.mock('@linode/api-v4/lib/account', async () => {
   const actual = await vi.importActual('@linode/api-v4/lib/account');
@@ -28,77 +20,79 @@ vi.mock('@linode/api-v4/lib/account', async () => {
   };
 });
 
-vi.mock('@tanstack/react-router', async () => {
-  const actual = await vi.importActual('@tanstack/react-router');
-  return {
-    ...actual,
-    useMatch: queryMocks.useMatch,
-  };
-});
-
 describe('BillingSummary', () => {
   it('displays appropriate helper text and value when there is no balance', async () => {
-    await renderWithThemeAndRouter(
+    renderWithTheme(
       <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
         <BillingSummary balance={0} balanceUninvoiced={5} paymentMethods={[]} />
-      </PayPalScriptProvider>
+      </PayPalScriptProvider>,
+      {
+        initialRoute: '/account/billing',
+      }
     );
     within(screen.getByTestId(accountBalanceText)).getByText(/no balance/i);
     within(screen.getByTestId(accountBalanceValue)).getByText('$0.00');
   });
 
   it('displays a credit when there is a negative balance', async () => {
-    await renderWithThemeAndRouter(
+    renderWithTheme(
       <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
         <BillingSummary
           balance={-10}
           balanceUninvoiced={5}
           paymentMethods={[]}
         />
-      </PayPalScriptProvider>
+      </PayPalScriptProvider>,
+      {
+        initialRoute: '/account/billing',
+      }
     );
     within(screen.getByTestId(accountBalanceText)).getByText(/credit/i);
     within(screen.getByTestId(accountBalanceValue)).getByText('$10.00');
   });
 
   it('displays the balance when there is a positive balance that is not yet past due', async () => {
-    await renderWithThemeAndRouter(
+    renderWithTheme(
       <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
         <BillingSummary
           balance={10}
           balanceUninvoiced={5}
           paymentMethods={[]}
         />
-      </PayPalScriptProvider>
+      </PayPalScriptProvider>,
+      {
+        initialRoute: '/account/billing',
+      }
     );
     within(screen.getByTestId(accountBalanceText)).getByText(/Balance/i);
     within(screen.getByTestId(accountBalanceValue)).getByText('$10.00');
   });
 
   it('does not display the promotions section unless there are promos', async () => {
-    const { rerender } = await renderWithThemeAndRouter(
+    const { rerender } = renderWithTheme(
       <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
         <BillingSummary balance={0} balanceUninvoiced={5} paymentMethods={[]} />
-      </PayPalScriptProvider>
+      </PayPalScriptProvider>,
+      {
+        initialRoute: '/account/billing',
+      }
     );
     expect(screen.queryByText('Promotions')).not.toBeInTheDocument();
     rerender(
-      wrapWithThemeAndRouter(
-        <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
-          <BillingSummary
-            balance={0}
-            balanceUninvoiced={5}
-            paymentMethods={[]}
-            promotions={promoFactory.buildList(1)}
-          />
-        </PayPalScriptProvider>
-      )
+      <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
+        <BillingSummary
+          balance={0}
+          balanceUninvoiced={5}
+          paymentMethods={[]}
+          promotions={promoFactory.buildList(1)}
+        />
+      </PayPalScriptProvider>
     );
     expect(screen.getByText('Promotions'));
   });
 
   it('renders promo summary, expiry, and credit remaining', async () => {
-    await renderWithThemeAndRouter(
+    renderWithTheme(
       <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
         <BillingSummary
           balance={0}
@@ -111,7 +105,10 @@ describe('BillingSummary', () => {
             summary: 'MY_PROMO_CODE',
           })}
         />
-      </PayPalScriptProvider>
+      </PayPalScriptProvider>,
+      {
+        initialRoute: '/account/billing',
+      }
     );
     const getByTextWithMarkup = withMarkup(screen.getByText);
     screen.getByText('MY_PROMO_CODE');
@@ -125,51 +122,45 @@ describe('BillingSummary', () => {
       promoFactory.build(),
       promoFactory.build({ service_type: 'linode', summary: 'MY_PROMO_CODE' }),
     ];
-    await renderWithThemeAndRouter(
+    renderWithTheme(
       <BillingSummary
         balance={0}
         balanceUninvoiced={5}
         paymentMethods={[]}
         promotions={promotions}
-      />
+      />,
+      {
+        initialRoute: '/account/billing',
+      }
     );
     expect(screen.queryByText('Applies to: All')).not.toBeInTheDocument();
     expect(screen.getByText('Applies to: Linodes'));
   });
 
   it('displays accrued charges', async () => {
-    await renderWithThemeAndRouter(
+    renderWithTheme(
       <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
         <BillingSummary balance={0} balanceUninvoiced={5} paymentMethods={[]} />
-      </PayPalScriptProvider>
+      </PayPalScriptProvider>,
+      {
+        initialRoute: '/account/billing',
+      }
     );
     within(screen.getByTestId('accrued-charges-value')).getByText('$5.00');
   });
 
   it('opens "Make a Payment" drawer when "Make a payment." is clicked', async () => {
-    const { getByTestId, getByText, rerender } = await renderWithThemeAndRouter(
+    const { getByTestId, getByText } = renderWithTheme(
       <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
         <BillingSummary balance={5} balanceUninvoiced={5} paymentMethods={[]} />
-      </PayPalScriptProvider>
+      </PayPalScriptProvider>,
+      {
+        initialRoute: '/account/billing',
+      }
     );
 
     const paymentButton = getByText('Make a payment', { exact: false });
     await userEvent.click(paymentButton);
-    queryMocks.useMatch.mockReturnValue({
-      routeId: '/account/billing/make-payment',
-    });
-
-    rerender(
-      wrapWithThemeAndRouter(
-        <PayPalScriptProvider options={{ clientId: PAYPAL_CLIENT_ID }}>
-          <BillingSummary
-            balance={5}
-            balanceUninvoiced={5}
-            paymentMethods={[]}
-          />
-        </PayPalScriptProvider>
-      )
-    );
 
     expect(getByTestId('drawer')).toBeVisible();
     expect(getByTestId('drawer-title').textContent).toEqual('Make a Payment');