+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeBackup/LinodeBackups.tsx
@@ -1,8 +1,6 @@
 import {
-  useGrants,
   useLinodeBackupsQuery,
   useLinodeQuery,
-  useProfile,
   useRegionsQuery,
   useTypeQuery,
 } from '@linode/queries';
@@ -25,8 +23,10 @@ import { TableCell } from 'src/components/TableCell';
 import { TableHead } from 'src/components/TableHead';
 import { TableRow } from 'src/components/TableRow';
 import { TableRowEmpty } from 'src/components/TableRowEmpty/TableRowEmpty';
+import { usePermissions } from 'src/features/IAM/hooks/usePermissions';
 import { getMonthlyBackupsPrice } from 'src/utilities/pricing/backups';
 
+import { useLinodeDetailContext } from '../LinodesDetailContext';
 import { BackupsPlaceholder } from './BackupsPlaceholder';
 import { BackupTableRow } from './BackupTableRow';
 import { CancelBackupsDialog } from './CancelBackupsDialog';
@@ -41,9 +41,18 @@ export const LinodeBackups = () => {
   const id = Number(linodeId);
 
   const navigate = useNavigate();
+  const { isBareMetalInstance } = useLinodeDetailContext();
 
-  const { data: profile } = useProfile();
-  const { data: grants } = useGrants();
+  const { permissions } = usePermissions(
+    'linode',
+    [
+      'cancel_linode_backups',
+      'create_linode_backup_snapshot',
+      'enable_linode_backups',
+    ],
+    linodeId
+  );
+
   const { data: linode } = useLinodeQuery(id);
   const { data: regions } = useRegionsQuery();
   const { data: type } = useTypeQuery(
@@ -55,12 +64,6 @@ export const LinodeBackups = () => {
     error,
     isLoading,
   } = useLinodeBackupsQuery(id, Boolean(linode?.backups.enabled));
-
-  const doesNotHavePermission =
-    Boolean(profile?.restricted) &&
-    grants?.linode.find(
-      (grant) => grant.id === linode?.id && grant.permissions !== 'read_write'
-    ) !== undefined;
 
   const [isRestoreDrawerOpen, setIsRestoreDrawerOpen] = React.useState(false);
 
@@ -102,11 +105,15 @@ export const LinodeBackups = () => {
     return (
       <BackupsPlaceholder
         backupsMonthlyPrice={backupsMonthlyPrice}
-        disabled={doesNotHavePermission}
+        disabled={!permissions.enable_linode_backups}
         linodeId={id}
         linodeIsInDistributedRegion={linodeIsInDistributedRegion}
       />
     );
+  }
+
+  if (isBareMetalInstance) {
+    return null;
   }
 
   if (error) {
@@ -146,34 +153,34 @@ export const LinodeBackups = () => {
                 {backups?.automatic.map((backup: LinodeBackup, idx: number) => (
                   <BackupTableRow
                     backup={backup}
-                    disabled={doesNotHavePermission}
                     handleDeploy={() => handleDeploy(backup)}
                     handleRestore={() => onRestoreBackup(backup)}
                     key={idx}
+                    linodeId={id}
                   />
                 ))}
                 {Boolean(backups?.snapshot.current) && (
                   <BackupTableRow
                     backup={backups!.snapshot.current!}
-                    disabled={doesNotHavePermission}
                     handleDeploy={() =>
                       handleDeploy(backups!.snapshot.current!)
                     }
                     handleRestore={() =>
                       onRestoreBackup(backups!.snapshot.current!)
                     }
+                    linodeId={id}
                   />
                 )}
                 {Boolean(backups?.snapshot.in_progress) && (
                   <BackupTableRow
                     backup={backups!.snapshot.in_progress!}
-                    disabled={doesNotHavePermission}
                     handleDeploy={() =>
                       handleDeploy(backups!.snapshot.in_progress!)
                     }
                     handleRestore={() =>
                       onRestoreBackup(backups!.snapshot.in_progress!)
                     }
+                    linodeId={id}
                   />
                 )}
               </>
@@ -186,13 +193,19 @@ export const LinodeBackups = () => {
           </TableBody>
         </Table>
       </Paper>
-      <CaptureSnapshot isReadOnly={doesNotHavePermission} linodeId={id} />
-      <ScheduleSettings isReadOnly={doesNotHavePermission} linodeId={id} />
+      <CaptureSnapshot
+        isReadOnly={!permissions.create_linode_backup_snapshot}
+        linodeId={id}
+      />
+      <ScheduleSettings
+        isReadOnly={!permissions.create_linode_backup_snapshot}
+        linodeId={id}
+      />
       <Box>
         <StyledCancelButton
           buttonType="outlined"
           data-qa-cancel
-          disabled={doesNotHavePermission}
+          disabled={!permissions.cancel_linode_backups}
           onClick={() => setIsCancelBackupsDialogOpen(true)}
         >
           Cancel Backups