+++ packages/manager/CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2025-07-21] - v1.146.2
+
+### Fixed:
+
+- Resort to payment id for edit billing flow ([#12544](https://github.com/linode/manager/pull/12544))
+
+## [2025-07-16] - v1.146.1
+
+
+### Fixed:
+
+- IAM RBAC: Fix a permission check for notification banner in Linode details component ([#12525](https://github.com/linode/manager/pull/12525))
+
 ## [2025-07-15] - v1.146.0